[Task 1849756] refactor MySQL connector UI components to share them with SQL server
