+++ PluginsAndFeatures/azure-toolkit-for-intellij/src/com/microsoft/azure/toolkit/intellij/connector/database/component/PasswordDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.intellij.connector.mysql.component.PasswordDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.microsoft.azure.toolkit.intellij.connector.database.component.PasswordDialog">
   <grid id="27dc6" binding="root" layout-manager="GridLayoutManager" row-count="6" column-count="4" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -72,7 +72,7 @@
           <text value="Save:"/>
         </properties>
       </component>
-      <component id="9530e" class="com.microsoft.azure.toolkit.intellij.connector.mysql.component.PasswordSaveComboBox" binding="passwordSaveComboBox">
+      <component id="9530e" class="com.microsoft.azure.toolkit.intellij.connector.database.component.PasswordSaveComboBox" binding="passwordSaveComboBox">
         <constraints>
           <grid row="1" column="3" row-span="1" col-span="1" vsize-policy="3" hsize-policy="0" anchor="0" fill="1" indent="0" use-parent-layout="false">
             <preferred-size width="110" height="30"/>
@@ -80,7 +80,7 @@
         </constraints>
         <properties/>
       </component>
-      <nested-form id="a0f94" form-file="com/microsoft/azure/toolkit/intellij/connector/mysql/component/TestConnectionActionPanel.form" binding="testConnectionActionPanel">
+      <nested-form id="a0f94" form-file="com/microsoft/azure/toolkit/intellij/connector/database/component/TestConnectionActionPanel.form" binding="testConnectionActionPanel">
         <constraints>
           <grid row="3" column="0" row-span="2" col-span="1" vsize-policy="2" hsize-policy="0" anchor="5" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
